Incorporated Alias indexing for metrics containing Wildcards
For wildcard metrics, to choose the meaningful name or remove the
unnecessary fields, need an index based aliasing.

For example, the metric query is springboot.gauge* and for the metric
springboot.gauge.hystrix.HystrixThreadPool.GP-API-TIERC-PRODUCT.reportingHosts,
need to display only last field 'reportingHosts' in the graph, use $6,
so that it can display the same. If it is $5$6 in alias field, it will
be GP-API-TIERC-PRODUCT.reportingHosts.

This modification is done in querybuilder.